+++ objects/stock.go
@@ -106,7 +106,7 @@ type StockQuote struct {
 	Eps                  *float64 `json:"eps"`
 	Pe                   *float64 `json:"pe"`
 	EarningsAnnouncement *string  `json:"earningsAnnouncement"` // 2020-07-22T16:09:24.000+0000
-	SharesOutstanding    *int64   `json:"sharesOutstanding"`
+	SharesOutstanding    *float64 `json:"sharesOutstanding"`
 	Timestamp            int64    `json:"timestamp"`
 }
